Working on perfecting the introduction page now
